requirements.txt
azure-functions
pymongo
azure-keyvault-keys
azure-identity